requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.7.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
ddt==1.2.1
DingtalkChatbot==1.3.0
Django==2.2.4
elasticsearch==7.0.2
et_xmlfile==1.0.1
Flask==1.1.1
futures==3.3.0
gevent==1.4.0
gitdb2==2.0.5
GitPython==2.1.13
greenlet==0.4.15
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
locust==0.0
locustio==0.11.0
lxml==4.4.0
MarkupSafe==1.1.1
more-itertools==7.2.0
msgpack==0.6.1
nose==1.3.7
numpy==1.17.0
openpyxl==2.6.2
paramiko==2.4.2
parsel==1.5.2
pluggy==0.12.0
py==1.8.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pymongo==3.8.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pytest==3.9.1
pytest-html==1.19.0
pytest-metadata==1.7.0
pytest-rerunfailures==4.2
python-crontab==2.3.8
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
pyzmq==17.1.2
queuelib==1.5.0
requests==2.19.1
Scrapy==1.7.3
selenium==3.141.0
service_identity==18.1.0
six==1.12.0
smmap2==2.0.5
sshtunnel==0.1.4
tomorrow==0.2.4
Twisted==19.7.0
urllib3==1.23
w3lib==1.20.0
Werkzeug==0.14.1
xlrd==1.1.0
you-get==0.4.1314
zope.interface==4.6.0
